extract operation support bundle route
